requirements.txt
bottle
PyYAML
docker-py
docker-compose==1.1.0-rc2
